+++ CHANGELOG.md
@@ -4,6 +4,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+
+- WASM bindings are exposed as a feature `bindings-wasm` in the main crate, to allow dependent crates to create their own WASM bindings and re-export some WASM-wrapped Umbral types. ([#78])
+
+
+[#78]: https://github.com/nucypher/rust-umbral/pull/78
+
+
 ## [0.5.0] - 2022-01-19
 
 ### Changed